add attendee & available partner service
