+++ TrashMob.Shared/Engine/EmailCopy/WelcomeToTrashMob.html
@@ -8,7 +8,7 @@
     Did you know you can <a href="https://www.trashmob.eco/locationpreference">set your location</a>? It's a great way to make sure you get notified when there are events created near you!
 </p>
 <p>
-    Ready to set up your first event? It's easy! Decide where you want to clean up, and <a href="https://www.trashmob.eco/manageeventdashboard">create it!</a>
+    Ready to set up your first event? It's easy! Decide where you want to clean up, and <a href="https://www.trashmob.eco/events/create">create it!</a>
 </p>
 <p>
     And don't forget to check out our <a href="https://www.trashmob.eco/gettingstarted">Getting Started</a> page for quick tips of what you can do to make your events a success!
